back-end/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
cloudinary==1.42.2
Django==5.1.6
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
dnspython==2.7.0
idna==3.10
mongoengine==0.29.1
pymongo==4.11.2
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0
